pyproject.toml
[project]
name = "gtask-mcp-server"
version = "0.1.0"
description = "MCP server to interact with Google Tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp[cli]>=1.15.0",
]
